clean up

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -1,33 +1,36 @@
 <?php
 
+use App\Enums\GenericStatus;
+use App\Models\Image;
+use App\Models\Role;
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-    /**
-     * Run the migrations.
-     */
     public function up(): void
     {
-        Schema::create('users', function (Blueprint $table) {
-            $table->id();
-            $table->string('name');
-            $table->string('email')->unique();
-            $table->boolean('is_superAdmin')->default(false);
-            $table->timestamp('email_verified_at')->nullable();
-            $table->string('password');
-            $table->rememberToken();
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('users')) {
+            Schema::create('users', function (Blueprint $table) {
+                $table->id();
+                $table->string('name');
+                $table->string('email')->unique();
+                $table->foreignIdFor(Role::class)->nullable();
+                $table->timestamp('email_verified_at')->nullable();
+                $table->string('password');
+                $table->foreignIdFor(Image::class)->nullable()->constrained();
+                $table->string('status')->nullable()->default(GenericStatus::enabled());
+                $table->rememberToken();
+                $table->timestamps();
+            });
+        }
     }
 
-    /**
-     * Reverse the migrations.
-     */
     public function down(): void
     {
-        Schema::dropIfExists('users');
+        if (Schema::hasTable('users')) {
+            Schema::dropIfExists('users');
+        }
     }
 };
