+++ database/migrations/2023_10_15_041134_create_blogs_table.php
@@ -1,29 +1,33 @@
 <?php
 
+use App\Enums\GenericStatus;
+use App\Models\Image;
+use App\Models\User;
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-    /**
-     * Run the migrations.
-     */
     public function up(): void
     {
-        Schema::create('blogs', function (Blueprint $table) {
-            $table->id();
-            $table->string('title');
-            $table->longText('body');
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('blogs')) {
+            Schema::create('blogs', function (Blueprint $table) {
+                $table->id();
+                $table->string('title');
+                $table->longText('body');
+                $table->foreignIdFor(Image::class)->nullable()->constrained();
+                $table->foreignIdFor(User::class)->nullable()->constrained();
+                $table->string('status')->nullable()->default(GenericStatus::enabled());
+                $table->timestamps();
+            });
+        }
     }
 
-    /**
-     * Reverse the migrations.
-     */
     public function down(): void
     {
-        Schema::dropIfExists('blogs');
+        if (Schema::hasTable('blogs')) {
+            Schema::dropIfExists('blogs');
+        }
     }
 };
