+++ database/migrations/2023_10_21_040957_create_directorates_table.php
@@ -1,29 +1,37 @@
 <?php
 
+use App\Enums\GenericStatus;
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-    /**
-     * Run the migrations.
-     */
     public function up(): void
     {
-        Schema::create('directorates', function (Blueprint $table) {
-            $table->id();
-            $table->string('title');
-            $table->longText('body');
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('directorates')) {
+            Schema::create('directorates', function (Blueprint $table) {
+                $table->id();
+                $table->string('title');
+                $table->longText('body');
+                $table->json('roles');
+                $table->string('directors_name');
+                $table->unsignedBigInteger('directors_image_id')->nullable();
+                $table->unsignedBigInteger('directorate_image_id')->nullable();
+                $table->string('status')->nullable()->default(GenericStatus::enabled());
+                $table->timestamps();
+
+                // foreign keys
+                $table->foreign('directors_image_id')->references('id')->on('images')->onDelete('set null');
+                $table->foreign('directorate_image_id')->references('id')->on('images')->onDelete('set null');
+            });
+        }
     }
 
-    /**
-     * Reverse the migrations.
-     */
     public function down(): void
     {
-        Schema::dropIfExists('directorates');
+        if (Schema::hasTable('directorates')) {
+            Schema::dropIfExists('directorates');
+        }
     }
 };
